Don't optimize floats to f32.
It's already the default for (python) floats to be taken as f32, so if
we encounter a different float, the user was intentional about it to
begin with, so we should respect that.

+++ plattli/writer.py
@@ -738,7 +738,7 @@ def _tightest_int(array):
 def _tight_dtype(array):
     array = np.asarray(array)
     if array.dtype.kind == "f":
-        return array.astype(np.float32, copy=False)
+        return array
     if array.dtype.kind in "iu":
         return _tightest_int(array)
     return None  # pragma: no cover - unreachable
